# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: A.N.O.T.H.E.R. Private API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /servers:
    get:
      operationId: listServers
      tags: [servers]
      summary: List servers the current user is a member of
      responses:
        "200":
          description: List of servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Server"
    post:
      operationId: createServer
      tags: [servers]
      summary: Create a new server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerInput"
      responses:
        "201":
          description: Server created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
  /servers/{serverId}:
    get:
      operationId: getServer
      tags: [servers]
      summary: Get server details
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Server details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
        "404":
          description: Server not found
    delete:
      operationId: deleteServer
      tags: [servers]
      summary: Delete a server (owner only)
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Server deleted
        "403":
          description: Forbidden
  /servers/{serverId}/join:
    post:
      operationId: joinServer
      tags: [servers]
      summary: Join a server
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerMember"
  /servers/{serverId}/leave:
    post:
      operationId: leaveServer
      tags: [servers]
      summary: Leave a server
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Left server
  /servers/{serverId}/members:
    get:
      operationId: getServerMembers
      tags: [servers]
      summary: Get all members of a server
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Server members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServerMember"
  /servers/{serverId}/members/{userId}/role:
    patch:
      operationId: updateMemberRole
      tags: [servers]
      summary: Update a member's role
      parameters:
        - name: serverId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberRoleUpdate"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerMember"
...
  /users/{userId}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get a user's profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found
  /users/me:
    patch:
      operationId: updateMyProfile
      tags: [users]
      summary: Update my profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, displayName, status, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [online, away, dnd, offline]
        role:
          type: string
          enum: [admin, member]
        createdAt:
          type: string
          format: date-time
    UserProfile:
      type: object
      required: [id, username, displayName, status, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [online, away, dnd, offline]
        role:
          type: string
          enum: [admin, member]
        createdAt:
          type: string
          format: date-time
...
    Server:
      type: object
      required: [id, name, ownerId, memberCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        iconUrl:
          type: ["string", "null"]
        ownerId:
          type: integer
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    ServerMember:
      type: object
      required: [id, serverId, userId, role, joinedAt, user]
      properties:
        id:
          type: integer
        serverId:
          type: integer
        userId:
          type: integer
        role:
          type: string
          enum: [owner, admin, member]
        joinedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/UserProfile"
...
    ServerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
...
    ProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 50
        bio:
          type: string
          maxLength: 200
        status:
          type: string
          enum: [online, away, dnd, offline]
    MemberRoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, member]
...