# lib/api-zod/src/generated/types/server.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * A.N.O.T.H.E.R. Private API
 * OpenAPI spec version: 0.1.0
 */

export interface Server {
  id: number;
  name: string;
  /** @nullable */
  iconUrl?: string | null;
  ownerId: number;
  memberCount: number;
  createdAt: Date;
}