# lib/api-zod/src/generated/types/serverInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * A.N.O.T.H.E.R. Private API
 * OpenAPI spec version: 0.1.0
 */

export interface ServerInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
}