# lib/api-zod/src/generated/types/serverMember.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * A.N.O.T.H.E.R. Private API
 * OpenAPI spec version: 0.1.0
 */
import type { ServerMemberRole } from './serverMemberRole';
import type { UserProfile } from './userProfile';

export interface ServerMember {
  id: number;
  serverId: number;
  userId: number;
  role: ServerMemberRole;
  joinedAt: Date;
  user: UserProfile;
}